Refactored Info component into sections
